+++ CHANGELOG.md
@@ -15,8 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Strict raft state. [#3167](https://github.com/evstack/ev-node/pull/3167)
 - Retry fetching the timestamp on error in da-client [#3166](https://github.com/evstack/ev-node/pull/3166)
 
-### Changes
+### Added
 
+- Add AWS KMS signer backend [#3171](https://github.com/evstack/ev-node/pull/3171)
 - Subscribe to forced inclusion namespace events [#3146](https://github.com/evstack/ev-node/pull/3146)
 
 ## v1.0.0